+++ packages/datepickers/src/index.ts
@@ -5,7 +5,7 @@
  * found at http://www.apache.org/licenses/LICENSE-2.0.
  */
 
-export { DatePicker } from './elements/DatePicker/DatePicker';
-export { DatePickerRange } from './elements/DatePickerRange/DatePickerRange';
+export { DatePicker } from './elements/Date_Picker/DatePicker';
+export { DatePickerRange } from './elements/Date_PickerRange/DatePickerRange';
 
 export type { IDatePickerProps, IDatePickerRangeProps } from './types';
